Use specific versions in package.json dependencies
Versions in dependencies should be specified rather than using wildcards
or URLs. If a dependency is updated to a new version with breaking
changes, this package might break as well. Pinning specific versions or
version ranges can almost ensure that breaking changes are not introduced.

This update uses [Caret Ranges][1] so that minor and patch versions are
updated, but major versions are not.

[1]: https://github.com/npm/node-semver/blob/v5.0.3/README.md#caret-ranges-123-025-004